requirements.txt
accelerate==1.10.1
antlr4-python3-runtime==4.9.3
certifi==2025.10.5
charset-normalizer==3.4.3
filelock==3.20.0
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
omegaconf==2.3.0
packaging==25.0
pandas==2.3.3
peft==0.17.1
psutil==7.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
